backend/requirement.txt
django
uvicorn[standard]
channels
channels_redis
djangorestframework
django-allauth